Added a base data storage
