+++ src/Happyr/LinkedIn/Storage/BaseDataStorage.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Happyr\LinkedIn\Storage;
+
+/**
+ * @author Tobias Nyholm
+ */
+abstract class BaseDataStorage implements DataStorageInterface
+{
+    public static $validKeys = array('state', 'code', 'access_token', 'user');
+
+    /**
+     * {@inheritDoc}
+     */
+    public function clearAll()
+    {
+        foreach (self::$validKeys as $key) {
+            $this->clear($key);
+        }
+    }
+
+    /**
+     * Generate a session name
+     *
+     * @param $key
+     *
+     * @return string
+     */
+    protected function constructSessionVariableName($key)
+    {
+        return 'linkedIn_'.$key;
+    }
+}
