+++ src/Happyr/LinkedIn/Storage/IlluminateSessionStorage.php
@@ -13,7 +13,7 @@ use Illuminate\Support\Facades\Session;
  * @author Andreas Creten
  *
  */
-class IlluminateSessionStorage extends SessionStorage
+class IlluminateSessionStorage extends BaseDataStorage
 {
     /**
      * {@inheritDoc}
